Start building packages when most downloads are finished
This change required a restructuring of the progress framework.
Instead of using rayon, we now use crossbeam directly and put the
packages in separate queues depending on which stage of the build
process they are in.  Once a package finishes one stage, it is
moved into the queue for the next stage.  This setup allows an
essentially seamless transition between two stages (unlike before,
where the user had to wait for all the downloads to finish in
order to begin the actual build process).